fix issue changing asset in the back cause crash

+++ remember/OpenResource/MJPhotoBrowser/MJPhoto.h
@@ -18,5 +18,5 @@
 
 // 是否已经保存到相册
 @property (nonatomic, assign) BOOL save;
-@property (nonatomic, assign) int index; // 索引
+@property (nonatomic, assign) NSInteger index; // 索引
 @end
